+++ app/code/community/Bitbull/Tooso/Helper/Data.php
@@ -142,6 +142,7 @@ class Bitbull_Tooso_Helper_Data extends Mage_Core_Helper_Abstract
         $client->setLogger(Mage::helper('tooso/log'));
         $client->setReportSender(Mage::helper('tooso/log_send'));
         $client->setSessionStorage(Mage::helper('tooso/session'));
+        $client->setAgent(Mage::helper('tooso/tracking')->getTrackingAgent());
 
         return $client;
     }
@@ -169,7 +170,7 @@ class Bitbull_Tooso_Helper_Data extends Mage_Core_Helper_Abstract
         $clientId = Mage::helper('tooso/session')->getClientId();
 
         $params = array(
-            'uip' => Mage::helper('core/http')->getRemoteAddr(),
+            'uip' => Mage::helper('tooso/tracking')->getRemoteAddr(),
             'ua' => Mage::helper('tooso/tracking')->getUserAgent(),
             'sessionId' => $sessionId,
             'cid' => $clientId,
@@ -222,5 +223,47 @@ class Bitbull_Tooso_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getUuid(){
         return Mage::helper('tooso')->getClient()->getUuid();
+    }
+
+    /**
+     * Build layout xml
+     * this is a fix to let Turpentine find block during ESI request
+     *
+     * @param $block Mage_Core_Block_Template
+     * @return string
+     */
+    public function addLayoutUpdate($block){
+        $layout = Mage::app()->getLayout();
+        $parent = $block->getParentBlock();
+        $referenceName = '';
+        if ($parent !== null) {
+            $referenceName = $block->getParentBlock()->getNameInLayout();
+        }
+        $layout->getUpdate()->addUpdate('
+            <reference name="'.$referenceName.'">
+                <block type="core/template" name="'.$block->getNameInLayout().'" template="" class="'.get_class($block).'"/>
+             </reference>
+        ');
+    }
+
+    /**
+     * Get current requested block for ESI request
+     *
+     * @return string
+     */
+    public function getESIRequestBlockName()
+    {
+        $req = Mage::app()->getRequest();
+        $esiHelper = Mage::helper('turpentine/esi');
+        $dataHelper = Mage::helper('turpentine/data');
+        if ($esiHelper === null || $dataHelper === null) {
+            return '';
+        }
+        $esiDataParamValue = $req->getParam( $esiHelper->getEsiDataParam() );
+        $esiDataArray = $dataHelper->thaw( $esiDataParamValue );
+        if (!isset($esiDataArray['name_in_layout'])) {
+            return '';
+        }
+        return $esiDataArray['name_in_layout'];
     }
 }