+++ app/code/community/Bitbull/Tooso/Helper/Tracking.php
@@ -44,7 +44,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getProductTrackingBlock($productId){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/tracking_productView');
+        $block = $layout->createBlock('tooso/tracking_productView', 'tooso_tracking_productView');
         $block->setProductID($productId);
         return $block;
     }
@@ -56,7 +56,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getPageTrackingBlock(){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/tracking_pageView');
+        $block = $layout->createBlock('tooso/tracking_pageView', 'tooso_tracking_pageView');
         return $block;
     }
 
@@ -68,7 +68,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getCheckoutTrackingBlock($orderId){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/tracking_checkout');
+        $block = $layout->createBlock('tooso/tracking_checkout', 'tooso_tracking_checkout');
         $block->setOrderId($orderId);
         return $block;
     }
@@ -80,7 +80,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getClearSearchIDBlock(){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/clearSearchId');
+        $block = $layout->createBlock('tooso/clearSearchId', 'tooso_clearSearchId');
         return $block;
     }
 
@@ -104,16 +104,12 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      * @return string
      */
     public function getRemoteAddr(){
-        if (!empty($_SERVER['HTTP_X_FORWARDED_FOR'])) {
-            $ips = explode(',', $_SERVER['HTTP_X_FORWARDED_FOR']);
-            return trim($ips[count($ips) - 1]);
-        } else {
-            if (!empty($_SERVER['HTTP_CLIENT_IP'])) {
-                return $_SERVER['HTTP_CLIENT_IP'];
-            } else {
-                return $_SERVER['REMOTE_ADDR'];
-            }
+        $remoteAddr = Mage::helper('core/http')->getRemoteAddr();
+        if (strpos($remoteAddr, ',') !== false){
+            $remoteAddrParts = explode(',', $remoteAddr);
+            $remoteAddr = $remoteAddrParts[0];
         }
+        return $remoteAddr;
     }
 
     /**
@@ -179,7 +175,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getTrackingLibraryBlock(){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/tracking_library');
+        $block = $layout->createBlock('tooso/tracking_library', 'tooso_tracking_library');
         return $block;
     }
 
@@ -190,7 +186,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getTrackingLibraryInitBlock(){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/tracking_libraryInit');
+        $block = $layout->createBlock('tooso/tracking_libraryInit', 'tooso_tracking_libraryInit');
         return $block;
     }
 
@@ -201,7 +197,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getPluginInfosBlock(){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/tracking_pluginInfos');
+        $block = $layout->createBlock('tooso/tracking_pluginInfos', 'tooso_tracking_pluginInfos');
         return $block;
     }
 
@@ -212,7 +208,7 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
      */
     public function getCustomerTrackingBlock(){
         $layout = Mage::app()->getLayout();
-        $block = $layout->createBlock('tooso/tracking_customerTracking');
+        $block = $layout->createBlock('tooso/tracking_customerTracking', 'tooso_tracking_customerTracking');
         return $block;
     }
 
@@ -267,7 +263,8 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
             'name' => $product->getName(),
             'brand' => $product->getBrand(),
             'price' => $product->getFinalPrice(),
-            'quantity' => 1
+            'quantity' => 1,
+            'position' => 0,
         ];
 
         $categoryIds = $product->getCategoryIds();
@@ -370,6 +367,19 @@ class Bitbull_Tooso_Helper_Tracking extends Mage_Core_Helper_Abstract
     public function isUserIdTrakingEnable($store = null)
     {
         return Mage::getStoreConfigFlag(self::XML_PATH_ANALYTICS_TRACK_USERID, $store);
+    }
+
+    /**
+     * Get tracking agent
+     *
+     * @return string
+     */
+    public function getTrackingAgent()
+    {
+        $agent = 'PHP/'.phpversion();
+        $agent .= ' Magento/'.Mage::getVersion();
+        $agent .= ' Tooso/'.Mage::getConfig()->getNode()->modules->Bitbull_Tooso->version;
+        return $agent;
     }
 
 }